backend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.3
Django==5.0.2
django-appconf==1.0.6
django-cors-headers==4.3.1
django-cryptography==1.1
django-phonenumber-field==8.0.0
django-qrcode==0.3
django-rest-auth==0.9.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.40
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
